% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-database.R
\name{build_database}
\alias{build_database}
\title{Build Database}
\usage{
build_database(override = FALSE)
}
\description{
Constructs a database from scratch by pulling all availabel items from Zotero.
}
