% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero.R
\name{build_id_string}
\alias{build_id_string}
\title{Build ID String}
\usage{
build_id_string(group = NA)
}
\arguments{
\item{group}{The group ID}

\item{user}{The user ID}
}
\value{
Returns a character variable with the user or group prefix
}
\description{
This function will determine if this the request is for a user library
or a group library and will return the appropriate code to be used to
access the Zotero API. This is an internal function that should not be
called directly by the user.
}
