% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-database.R
\name{convert_keys_to_list_of_lists}
\alias{convert_keys_to_list_of_lists}
\title{Convert Keys to List of Lists}
\usage{
convert_keys_to_list_of_lists(kl)
}
\arguments{
\item{kl}{A character vector of keys of the form generated by `get_zotero_key_list`}
}
\value{
A list of character vectors of keys where no item in the list is more
than 50 elements long.
}
\description{
Converts a character vector of keys into a list of character vectors of keys.
Since only 50 items can be read from Zotero at a time, this function converts
a list of keys into a series of character vectors that are each no more than
50 items long.
}
\details{
DEPRECATED:  Pulling more than one record at a time does not seem feasible as
many of the keys do not return a BibLatex record and there is no way to tell
which keys returned nothing when more than one key is read at a time. So this
function will likely be unnecessary.
}
