% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero.R
\name{get_zotero_key}
\alias{get_zotero_key}
\title{Get Zotero Key}
\usage{
get_zotero_key()
}
\value{
Character variable containing the Zotero API key.
}
\description{
This function retrieves the Zotero key or throws an error.
}
