% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero.R
\name{list_zotero_collections}
\alias{list_zotero_collections}
\title{List Zotero Collections}
\usage{
list_zotero_collections(group = NA)
}
\arguments{
\item{group}{Group ID (default = NA)}
}
\value{
A named vector that with the text-based names of each collection and
the key value for that collection that is used to access the API.
}
\description{
This function retrieves a list of all collections associated with a user
or group.
}
