% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-database.R
\name{pull_new_items_from_zotero}
\alias{pull_new_items_from_zotero}
\title{Pull New Keys from Zotero}
\usage{
pull_new_items_from_zotero(key_list)
}
\arguments{
\item{key_list}{Character vector of keys, such as that produced by `get_zotero_key_list`.}
}
\value{
Data.frame of BibLatex records
}
\description{
Pulls the Biblatex records from Zotero from the list of keys supplied to the
function.
}
