% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-database.R
\name{read_citations}
\alias{read_citations}
\title{Read Citations from RMD File}
\usage{
read_citations(rmd_file)
}
\arguments{
\item{rmd_file}{Path to the R Markdown file to be read}
}
\value{
A character vector of citations (excluding the starting at symbol)
}
\description{
Reads a R Markdown file and generates a list of the citations that are contained
within it. Note that the function only reads a single Rmd file and does not
currently address files that might be linked to in the Rmd file or citations
that only appear in a table or figure.
}
