% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-database.R
\name{remove_blanks}
\alias{remove_blanks}
\title{Remove Blanks}
\usage{
remove_blanks(x)
}
\arguments{
\item{x}{A character vector}
}
\value{
A character vector with any blank elements removed
}
\description{
Removes blank elements from a character vector.
}
