% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-database.R
\name{retrieve_single_record_from_zotero}
\alias{retrieve_single_record_from_zotero}
\title{Retrieve Single Record from Zotero}
\usage{
retrieve_single_record_from_zotero(key)
}
\arguments{
\item{key}{Character vector scalar giving the key for a single Zotero record.  Equivalent
to one element from `get_zotero_key_list`.}
}
\value{
A data.frame with a single record containing, the key, the name of the record,
and the BibLatex record text.
}
\description{
Retrieves a the Biblatex for a single item on Zotero.
}
