% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-database.R
\name{update_zotero_database}
\alias{update_zotero_database}
\title{Update Zotero Database}
\usage{
update_zotero_database()
}
\description{
Updates an existing Zotero database.  Reads in the database and reads in
all keys that have been changed since the last update.
}
