% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero.R
\name{validate_collections}
\alias{validate_collections}
\title{Validate Collections}
\usage{
validate_collections(group = NA, collections)
}
\value{
A named character vector with the IDs of each collection found in the
online database
}
\description{
This function takes the collections provided by the user and validates that
they exist online. If none of them exist, an error will be thrown.  If some
of them do not exist, a warning will be issued.
}
